Refactor tx params
