+++ src/routes/safe/components/Balances/SendModal/screens/ReviewSendFundsTx/index.tsx
@@ -25,19 +25,13 @@ import { SpendingLimit } from 'src/logic/safe/store/models/safe'
 import { sameString } from 'src/utils/strings'
 import { TokenProps } from 'src/logic/tokens/store/model/token'
 import { RecordOf } from 'immutable'
-import { EstimationStatus, useEstimateTransactionGas } from 'src/logic/hooks/useEstimateTransactionGas'
-import { useEstimationStatus } from 'src/logic/hooks/useEstimationStatus'
-import { ButtonStatus, Modal } from 'src/components/Modal'
-import { ReviewInfoText } from 'src/components/ReviewInfoText'
 
 import { styles } from './style'
-import { EditableTxParameters } from 'src/routes/safe/components/Transactions/helpers/EditableTxParameters'
-import { TxParametersDetail } from 'src/routes/safe/components/Transactions/helpers/TxParametersDetail'
+import { TxParamsState } from 'src/routes/safe/components/Transactions/helpers/TxParamsState'
 import { TxParameters } from 'src/routes/safe/container/hooks/useTransactionParameters'
 import { Errors, logError } from 'src/logic/exceptions/CodedException'
 import { ModalHeader } from '../ModalHeader'
 import { extractSafeAddress } from 'src/routes/routes'
-import ExecuteCheckbox from 'src/components/ExecuteCheckbox'
 import { getNativeCurrencyAddress } from 'src/config/utils'
 
 const useStyles = makeStyles(styles)
@@ -97,45 +91,11 @@ const ReviewSendFundsTx = ({ onClose, onPrev, tx }: ReviewTxProps): React.ReactE
   const isSendingNativeToken = useMemo(() => sameAddress(txToken?.address, getNativeCurrencyAddress()), [txToken])
   const txRecipient = isSendingNativeToken ? tx.recipientAddress : txToken?.address || ''
   const txValue = isSendingNativeToken ? toTokenUnit(tx.amount, nativeCurrency.decimals) : '0'
-  const data = useTxData(isSendingNativeToken, tx.amount, tx.recipientAddress, txToken)
-  const [manualSafeTxGas, setManualSafeTxGas] = useState('0')
-  const [manualGasPrice, setManualGasPrice] = useState<string | undefined>()
-  const [manualGasLimit, setManualGasLimit] = useState<string | undefined>()
+  const txData = useTxData(isSendingNativeToken, tx.amount, tx.recipientAddress, txToken)
 
-  const {
-    gasCostFormatted,
-    gasPriceFormatted,
-    gasMaxPrioFeeFormatted,
-    gasLimit,
-    gasEstimation,
-    txEstimationExecutionStatus,
-    isExecution,
-    isCreation,
-    isOffChainSignature,
-  } = useEstimateTransactionGas({
-    txData: data,
-    txRecipient,
-    txType: tx.txType,
-    txAmount: txValue,
-    safeTxGas: manualSafeTxGas,
-    manualGasPrice,
-    manualGasLimit,
-  })
-
-  const [buttonStatus, setButtonStatus] = useEstimationStatus(txEstimationExecutionStatus)
   const isSpendingLimit = sameString(tx.txType, 'spendingLimit')
-  const [executionApproved, setExecutionApproved] = useState<boolean>(true)
-  const doExecute = isExecution && executionApproved
 
-  const submitTx = async (txParameters: TxParameters) => {
-    setButtonStatus(ButtonStatus.LOADING)
-
-    if (!safeAddress) {
-      setButtonStatus(ButtonStatus.READY)
-      logError(Errors._802)
-      return
-    }
-
+  const submitTx = async (txParameters: TxParameters, delayExecution: boolean) => {
     if (isSpendingLimit && txToken && tx.tokenSpendingLimit) {
       const spendingLimitTokenAddress = isSendingNativeToken ? ZERO_ADDRESS : txToken.address
       const spendingLimit = getSpendingLimitContract()
@@ -154,7 +114,6 @@ const ReviewSendFundsTx = ({ onClose, onPrev, tx }: ReviewTxProps): React.ReactE
           .send({ from: tx.tokenSpendingLimit.delegate })
           .on('transactionHash', () => onClose())
       } catch (err) {
-        setButtonStatus(ButtonStatus.READY)
         logError(Errors._801, err.message)
       }
       return
@@ -165,143 +124,68 @@ const ReviewSendFundsTx = ({ onClose, onPrev, tx }: ReviewTxProps): React.ReactE
         safeAddress: safeAddress,
         to: txRecipient as string,
         valueInWei: txValue,
-        txData: data,
+        txData,
         txNonce: txParameters.safeNonce,
         safeTxGas: txParameters.safeTxGas,
         ethParameters: txParameters,
         notifiedTransaction: TX_NOTIFICATION_TYPES.STANDARD_TX,
-        delayExecution: !executionApproved,
+        delayExecution,
       }),
     )
     onClose()
   }
 
-  const closeEditModalCallback = (txParameters: TxParameters) => {
-    const oldGasPrice = gasPriceFormatted
-    const newGasPrice = txParameters.ethGasPrice
-    const oldMaxPrioFee = gasMaxPrioFeeFormatted
-    const newMaxPrioFee = txParameters.ethMaxPrioFee
-    const oldSafeTxGas = gasEstimation
-    const newSafeTxGas = txParameters.safeTxGas
-
-    if (newGasPrice && oldGasPrice !== newGasPrice) {
-      setManualGasPrice(txParameters.ethGasPrice)
-    }
-
-    if (newMaxPrioFee && oldMaxPrioFee !== newMaxPrioFee) {
-      setManualGasPrice(txParameters.ethMaxPrioFee)
-    }
-
-    if (txParameters.ethGasLimit && gasLimit !== txParameters.ethGasLimit) {
-      setManualGasLimit(txParameters.ethGasLimit)
-    }
-
-    if (newSafeTxGas && oldSafeTxGas !== newSafeTxGas) {
-      setManualSafeTxGas(newSafeTxGas)
-    }
-  }
-
   return (
-    <EditableTxParameters
-      isOffChainSignature={isOffChainSignature}
-      isExecution={doExecute}
-      ethGasLimit={gasLimit}
-      ethGasPrice={gasPriceFormatted}
-      ethMaxPrioFee={gasMaxPrioFeeFormatted}
-      safeTxGas={gasEstimation}
-      closeEditModalCallback={closeEditModalCallback}
-    >
-      {(txParameters, toggleEditMode) => (
-        <>
-          {/* Header */}
-          <ModalHeader onClose={onClose} subTitle="2 of 2" title="Send funds" />
+    <TxParamsState txData={txData} txValue={txValue} txType={tx.txType || ''} onSubmit={submitTx} onBack={onPrev}>
+      <>
+        {/* Header */}
+        <ModalHeader onClose={onClose} subTitle="2 of 2" title="Send funds" />
 
-          <Hairline />
+        <Hairline />
 
-          <Block className={classes.container}>
-            {/* SafeInfo */}
-            <SafeInfo />
-            <Divider withArrow />
+        <Block className={classes.container}>
+          {/* SafeInfo */}
+          <SafeInfo />
+          <Divider withArrow />
 
-            {/* Recipient */}
-            <Row margin="xs">
-              <Paragraph color="disabled" noMargin size="md" style={{ letterSpacing: '-0.5px' }}>
-                Recipient
-              </Paragraph>
-            </Row>
-            <Row align="center" margin="md" data-testid="recipient-review-step">
-              <Col xs={12}>
-                <PrefixedEthHashInfo
-                  hash={tx.recipientAddress}
-                  name={tx.recipientName}
-                  showCopyBtn
-                  showAvatar
-                  explorerUrl={getExplorerInfo(tx.recipientAddress)}
-                />
-              </Col>
-            </Row>
-
-            {/* Amount */}
-            <Row margin="xs">
-              <Paragraph color="disabled" noMargin size="md" style={{ letterSpacing: '-0.5px' }}>
-                Amount
-              </Paragraph>
-            </Row>
-            <Row align="center" margin="md">
-              <Img alt={txToken?.name as string} height={28} onError={setImageToPlaceholder} src={txToken?.logoUri} />
-              <Paragraph
-                className={classes.amount}
-                noMargin
-                size="md"
-                data-testid={`amount-${txToken?.symbol as string}-review-step`}
-              >
-                {tx.amount} {txToken?.symbol}
-              </Paragraph>
-            </Row>
-
-            {isExecution && !isSpendingLimit && <ExecuteCheckbox onChange={setExecutionApproved} />}
-
-            {/* Tx Parameters */}
-            {/* FIXME TxParameters should be updated to be used with spending limits */}
-            {!isSpendingLimit && (
-              <TxParametersDetail
-                txParameters={txParameters}
-                onEdit={toggleEditMode}
-                isTransactionCreation={isCreation}
-                isTransactionExecution={doExecute}
-                isOffChainSignature={isOffChainSignature}
+          {/* Recipient */}
+          <Row margin="xs">
+            <Paragraph color="disabled" noMargin size="md" style={{ letterSpacing: '-0.5px' }}>
+              Recipient
+            </Paragraph>
+          </Row>
+          <Row align="center" margin="md" data-testid="recipient-review-step">
+            <Col xs={12}>
+              <PrefixedEthHashInfo
+                hash={tx.recipientAddress}
+                name={tx.recipientName}
+                showCopyBtn
+                showAvatar
+                explorerUrl={getExplorerInfo(tx.recipientAddress)}
               />
-            )}
-          </Block>
+            </Col>
+          </Row>
 
-          {/* Disclaimer */}
-          {/* FIXME Estimation should be fixed to be used with spending limits */}
-          {!isSpendingLimit && txEstimationExecutionStatus !== EstimationStatus.LOADING && (
-            <ReviewInfoText
-              gasCostFormatted={gasCostFormatted}
-              isCreation={isCreation}
-              isExecution={doExecute}
-              isOffChainSignature={isOffChainSignature}
-              safeNonce={txParameters.safeNonce}
-              txEstimationExecutionStatus={txEstimationExecutionStatus}
-            />
-          )}
-
-          {/* Footer */}
-          <Modal.Footer withoutBorder={!isSpendingLimit && buttonStatus !== ButtonStatus.LOADING}>
-            <Modal.Footer.Buttons
-              cancelButtonProps={{ onClick: onPrev, text: 'Back' }}
-              confirmButtonProps={{
-                onClick: () => submitTx(txParameters),
-                status: buttonStatus,
-                text: txEstimationExecutionStatus === EstimationStatus.LOADING ? 'Estimating' : undefined,
-                testId: 'submit-tx-btn',
-              }}
-            />
-          </Modal.Footer>
-        </>
-      )}
-    </EditableTxParameters>
+          {/* Amount */}
+          <Row margin="xs">
+            <Paragraph color="disabled" noMargin size="md" style={{ letterSpacing: '-0.5px' }}>
+              Amount
+            </Paragraph>
+          </Row>
+          <Row align="center" margin="md">
+            <Img alt={txToken?.name as string} height={28} onError={setImageToPlaceholder} src={txToken?.logoUri} />
+            <Paragraph
+              className={classes.amount}
+              noMargin
+              size="md"
+              data-testid={`amount-${txToken?.symbol as string}-review-step`}
+            >
+              {tx.amount} {txToken?.symbol}
+            </Paragraph>
+          </Row>
+        </Block>
+      </>
+    </TxParamsState>
   )
 }
